Create TextInput component to simplify App.js
